refactor: renomeia ResultType para TransactionResult e Discount para float
- Renomeia classe ResultType para TransactionResult (mais semântico)
- Altera Query.Discount de int? para float? (suporta casas decimais)
- Atualiza todas as referências no código e testes

+++ BookPricesWatcher.Business/Interfaces/IQueryHistoryService.cs
@@ -79,6 +79,6 @@ public class QueryDetailDto
     public string? Title { get; set; }
     public string? Author { get; set; }
     public decimal? Price { get; set; }
-    public int? Discount { get; set; }
+    public float? Discount { get; set; }
     public string? ErrorMessage { get; set; }
 }
